requirements.txt
gspread==0.2.2
oauthlib==0.7.2
python-twitter==2.2
requests==2.5.3
requests-oauthlib==0.4.2
